Crate nosqlite

This crate implements a NoSQL-like API over SQLite using SQLite's Json1 extension.

If this is your first time here, start off by reading the Connection docs.

Modules

util
A module for utility structs that don't do much on their own

Macros

json
Construct a serde_json::Value from a JSON literal.

Structs

Column
A column in the SQL table.

Connection
A connection the underlying sqlite database.

Entry
Represents one 'row' of the table.

Field
Represents a field in a JSON object.

Iterator
Represents a potential operation on a table.

Json
A newtype to implement the [ToSql] and [FromSql] traits for a struct that implements [Serialize] and [Deserialize] respectively.

Operation
Represents an operation to get a JSON object using its id key.

Table
A table in the database.

Traits

Filter
Represents a condition which will determine what entries the operation can work on.

FromSql
A trait for types that can be created from a SQLite value.

Key
This can be used for filters or getting fields

Sort
An expression that is used to sort the entries in the database before it is returned to the program.

ToSql
A trait for types that can be converted into SQLite values.

Functions

column
Create a representation of a column in a SQL table.

field
Creates a representation of a field in a JSON object.