Struct nosqlite::Iterator

#[must_use = "This struct does not do anything until executed"]
pub struct Iterator<'a, I, W, S> { /* fields omitted */ }

Represents a potential operation on a table.

Methods

impl<'a, I: FromSql, W: Filter, S: Sort> Iterator<'a, I, W, S>

pub fn data<T: DeserializeOwned, C: AsRef<SqliteConnection>>(
&self,
connection: C
) -> SqliteResult<Vec<T>>

GETs only the JSON object.

Example

let people: Vec<Person> = table.iter().data(&connection)?;

pub fn entry<T: DeserializeOwned, C: AsRef<SqliteConnection>>(
&self,
connection: C
) -> SqliteResult<Vec<Entry<I, T>>>

GETs the id and the JSON object.

Example

let people: Vec<Entry<i64, Person>> = table.iter().entry(&connection)?;

pub fn id<C: AsRef<SqliteConnection>>(
&self,
connection: C
) -> SqliteResult<Vec<I>>

GETs just the id of the entry.

Example

let people: Vec<i64> = table.iter().id(&connection)?;

pub fn field<T: FromSql, C: AsRef<SqliteConnection>>(
&self,
field_: &str,
connection: C
) -> SqliteResult<Vec<T>>

GETs a field of the JSON object.

Example

#[derive(Deserialize, Serialize)]
struct Person {
	first_name: String,
	last_name: String,
	age: u8,
}
table.insert(Person{ first_name: "Hiruna".into(), last_name: "Jayamanne".into(), age: 19 }, &connection)?;
let people: Vec<String> = table.iter().field("first_name", &connection)?;
assert_eq!(people[0], "Hiruna");

pub fn fields<T, F, C, A>(
&self,
fields: F,
connection: C
) -> SqliteResult<Vec<T>> where
F: IntoIterator<Item = A>,
A: AsRef<str>,
T: DeserializeOwned,
C: AsRef<SqliteConnection>,

GETs multiple fields from the JSON object.

Example

#[derive(Deserialize, Serialize)]
struct Person {
	first_name: String,
	last_name: String,
	age: u8,
}
table.insert(Person{ first_name: "Hiruna".into(), last_name: "Jayamanne".into(), age: 19 }, &connection)?;
let people: Vec<(String, String)> = table.iter().fields(&["first_name", "last_name"], &connection)?;
assert_eq!(people[0], ("Hiruna".into(), "Jayamanne".into()));

pub fn insert<T, C>(
&self,
field: &str,
value: T,
connection: C
) -> SqliteResult<()> where
T: ToSql,
C: AsRef<SqliteConnection>,

Inserts a field into the JSON object with a given value.

If the field already exists, nothing will happen. If you wish for it to be overwritten, use set instead.

Example

// Has an `age` field
table.insert(json!({ "first_name": "Hiruna", "age": 19 }), &connection)?;
// Does not have an `age` field
table.insert(json!({ "first_name": "Bob" }), &connection)?;
// `insert` should only work for objects which don't have the field already
table.iter().insert("age", 13, &connection);
let people: Vec<(String, u8)> = table.iter().fields(&["first_name", "age"], &connection)?;
assert_eq!(people[0], ("Hiruna".into(), 19));
assert_eq!(people[1], ("Bob".into(), 13)); // Only Bob was changed

pub fn patch<T, C>(&self, value: T, connection: C) -> SqliteResult<()> where
T: Serialize,
C: AsRef<SqliteConnection>,

Uses a JSON object update or create fields in the entry's JSON object.

Any fields that do not exist will be created. This unfortunately does not work with Arrays. It will replace Arrays instead. To insert into Arrays, use insert instead.

Example

table.insert(json!({ "first_name": "Hiruna", "age": 19 }), &connection)?;
table.insert(json!({ "first_name": "Bob" }), &connection)?;
table.insert(json!({ "first_name": "Alex", "grades": [6, 8, 7] }), &connection)?;
// `patch` overwrites any field
// arrays are completely replace
table.iter().patch(json!({ "age": 13, "grades": [9] }), &connection);
let people: Vec<(u8, Vec<u8>)> = table.iter().fields(&["age", "grades"], &connection)?;
for person in people.into_iter() {
	// `age` field was overwritten and set to 13
	assert_eq!(person.0, 13);
	// `grades` field was overwritten and set to an array of one element
	assert_eq!(person.1, [9])
}

pub fn remove<C>(&self, field: &str, connection: C) -> SqliteResult<()> where
C: AsRef<SqliteConnection>,

Removes a field from a JSON object.

Example

table.insert(json!({ "first_name": "Hiruna", "age": 19 }), &connection)?;
table.insert(json!({ "first_name": "Bob", "age": 13 }), &connection)?;

table.iter().remove("age", &connection);
// This should return no entries because no JSON object has an `age` field
let people: Vec<u8> = table.iter().field("age", &connection)?;
assert_eq!(people.len(), 0);
// This *does not* delete the entries
assert_eq!(table.iter().id(&connection)?.len(), 2);

pub fn replace<T, C>(
&self,
field: &str,
value: T,
connection: C
) -> SqliteResult<()> where
T: ToSql,
C: AsRef<SqliteConnection>,

Replaces a field in a JSON object with a given value.

Will only replace an already existing field.

Example

// Has an `age` field
table.insert(json!({ "first_name": "Hiruna", "age": 19 }), &connection)?;
// Does not have an `age` field
table.insert(json!({ "first_name": "Bob" }), &connection)?;
// `replace` should only work for objects which have the field already
table.iter().replace("age", 13, &connection);
let people: Vec<u8> = table.iter().field("age", &connection)?;
// Both objects had their fields set
assert_eq!(people.len(), 1);
assert_eq!(people[0], 13);

pub fn set<T, C>(
&self,
field: &str,
value: T,
connection: C
) -> SqliteResult<()> where
T: ToSql,
C: AsRef<SqliteConnection>,

Sets a field in a JSON object to a given field.

If the field does not exist, it will be created. If the field does already exist, it will be overwritten. If you wish for the value to not be overwritten, use insert instead.

Example

// Has an `age` field
table.insert(json!({ "first_name": "Hiruna", "age": 19 }), &connection)?;
// Does not have an `age` field
table.insert(json!({ "first_name": "Bob" }), &connection)?;
table.iter().set("age", 13, &connection);
let people: Vec<u8> = table.iter().field("age", &connection)?;
// Both objects had their fields set
assert_eq!(people[0], 13);
assert_eq!(people[1], 13);

pub fn delete<C: AsRef<SqliteConnection>>(
&self,
connection: C
) -> SqliteResult<()>

Deletes the entry.

Example

// Add some random numbers to the table
table.insert(9, &connection)?;
table.insert(10, &connection)?;
table.insert(12, &connection)?;
table.insert(3, &connection)?;
// We expect 4 entries
let length = table.iter().id(&connection)?.len();
assert_eq!(length, 4);
// Now we'll remove every number 10 or above.
table.iter().filter(field("").gte(10)).delete(&connection)?;
// There should only be 2 entries left.
let length = table.iter().id(&connection)?.len();
assert_eq!(length, 2);

pub fn filter<A: Filter>(self, filter: A) -> Iterator<'a, I, A, S>

Applies a filter on what entries the command will operate on.

Example

table.insert(json!({"name": "Hiruna", "age": 19}), &connection)?;
table.insert(json!({"name": "Bob", "age": 13}), &connection)?;
table.insert(json!({"name": "Callum", "age": 12}), &connection)?;
table.insert(json!({"name": "John", "age": 20}), &connection)?;
// We only want people 18 or above
let people: Vec<(String, u8)> = table.iter()
	.filter(field("age").gte(18))
	.fields(&["name", "age"], &connection)?;
// Only 2 people should have passed the filter
assert_eq!(people.len(), 2);
for person in people.into_iter() {
	// They should be 18+ years old
	assert!(person.1 >= 18);
}

pub fn sort<A: Sort>(self, key: A) -> Iterator<'a, I, W, A>

Sort by using a specific field(s)

Examples

table.insert(json!({"a": 3, "b": 2}), &connection)?;
table.insert(json!({"a": 6, "b": 1}), &connection)?;
table.insert(json!({"a": 2, "b": 8}), &connection)?;
table.insert(json!({"a": 8, "b": 4}), &connection)?;

let data: Vec<u8> = table.iter().sort(field("a").ascending()).field("a", &connection)?;

assert_eq!(data[0], 2);
assert_eq!(data[1], 3);
assert_eq!(data[2], 6);
assert_eq!(data[3], 8);

table.insert(json!({"a": 2, "b": 2}), &connection)?;
table.insert(json!({"a": 2, "b": 1}), &connection)?;
table.insert(json!({"a": 8, "b": 4}), &connection)?;
table.insert(json!({"a": 2, "b": 8}), &connection)?;

let data: Vec<(u8, u8)> = table.iter()
	.sort(field("a").ascending().and(field("b").ascending()))
	.fields(&["a","b"], &connection)?;

assert!(data[0].0 == 2 && data[0].1 == 1);
assert!(data[1].0 == 2 && data[1].1 == 2);
assert!(data[2].0 == 2 && data[2].1 == 8);
assert!(data[3].0 == 8 && data[3].1 == 4);

pub fn skip(self, n: u32) -> Self

Skip over n entries.

Example

table.insert(json!({"name": "Hiruna", "age": 19}), &connection)?;
table.insert(json!({"name": "Bob", "age": 13}), &connection)?;
table.insert(json!({"name": "Callum", "age": 12}), &connection)?;
table.insert(json!({"name": "John", "age": 20}), &connection)?;
// We want the people third or above in position within the table
let people: Vec<String> = table.iter()
	.skip(2)
	.field("name", &connection)?;
// Only 2 people should have passed the filter
assert_eq!(people.len(), 2);
assert_eq!(people[0], "Callum");
assert_eq!(people[1], "John");

pub fn take(self, n: u32) -> Self

Take only n entries.

Example

table.insert(json!({"name": "Hiruna", "age": 19}), &connection)?;
table.insert(json!({"name": "Bob", "age": 13}), &connection)?;
table.insert(json!({"name": "Callum", "age": 12}), &connection)?;
table.insert(json!({"name": "John", "age": 20}), &connection)?;
// We only want the first two people
let people: Vec<String> = table.iter()
	.take(2)
	.field("name", &connection)?;
// Only 2 people should have passed the filter
assert_eq!(people.len(), 2);
assert_eq!(people[0], "Hiruna");
assert_eq!(people[1], "Bob");

pub fn execute<A, F, C>(
&self,
command: &str,
execute: F,
connection: C
) -> SqliteResult<A> where
F: FnOnce(Statement, Vec<(&str, &dyn ToSql)>) -> SqliteResult<A>,
C: AsRef<SqliteConnection>,

Execute a query using the given command (e.g. "SELECT data"), the given function to handle the output, and the connection to the database.

It is not recommended to use this method.