Struct nosqlite::Table

pub struct Table<I> {
    pub id: String,
    pub data: String,
    pub name: String,
    // some fields omitted
}

A table in the database.

Use [Connection::table] to make this struct but if you have a custom schema, you should make this struct by providing the fields your self.

The generic type parameter, I, is the type of the Id column.

Fields
id: String

The id column.

The default implementation uses an integer but the value does not matter. e.g. a String UUID can be used.

data: String

The column that stores the JSON object.

The default implementation uses a Non Null Text but the value can be nullable or a Blob.

name: String

The name of the table.

Methods

impl<A> Table<A>

pub fn index<S, I, T, C>(
&self,
name: S,
keys: I,
connection: C
) -> SqliteResult<()> where
S: Display,
I: IntoIterator<Item = T>,
T: Key,
C: AsRef<SqliteConnection>,

Creates an index on the table with the given keys.

This is meant to speed up queries but whether it is actually used or not is determined by SQLite at runtime so I would suggest running some sort of benchmark to see if creating the index actually does speed up query times.

Example

table.index("my_index", &[field("name"), field("age")], connection)?;

If you want to index both a field and a column then you need to cast the reference to a &dyn Key.

table.index("my_index", &[&field("name") as &dyn Key, &column("id") as &dyn Key], connection)?;

impl<I: FromSql> Table<I>

pub fn unchecked<K: FromSql, T: Into<String>>(id: T, data: T, name: T) -> Self

Creates a table but doesn't check if the table exists.

This can be used to give a custom name and type of the id and data columns.

Generally using Connection::table is recommended instead of this.

pub fn iter(&self) -> Iterator<I, (), ()>

Iterate through all the entries in the table.

Example

let table = connection.table("people")?;
table.insert(json!({"name": "Hiruna", "age": 19}), &connection)?;
table.insert(json!({"name": "Bobby", "age": 13}), &connection)?;
let data = table.iter()
	.filter(field("age").gte(18))
	.fields::<(String, u8), _, _, _>(&["name", "age"], &connection)?;
assert_eq!(data.len(), 1);
assert!(data[0].1 > 18);

pub fn insert<T: Serialize, C: AsRef<SqliteConnection>>(
&self,
data: T,
connection: C
) -> SqliteResult<()>

Inserts a JSON object into the data column of the table.

Multiple JSON objects that are exactly the same can be inserted.

Warning: If your table has other columns that are not nullable, then you should not use this.

Example

// This table is empty
let table = connection.table("people")?;
// We'll insert two JSON objects which are exactly the same
let json = json!({"name": "Hiruna", "age": 19});
table.insert(&json, &connection)?;
table.insert(&json, &connection)?;
// Now we'll check how many entries the table has
let length = table.iter().id(&connection)?.len();
assert_eq!(length, 2);

impl<I: FromSql + ToSql> Table<I>

pub fn get(&self, id: I) -> Operation<I>

Gets a JSON object using a id from the id column.

Example

#[derive(Deserialize, Serialize)]
struct Person {
	name: String,
}
// The table of these ids are integers
let table = connection.table("people")?;
table.insert(&Person {name: "Hiruna".into()}, &connection)?;
table.insert(&Person {name: "Bobby".into()}, &connection)?;
// Now we'll get the 2nd entry from the table.
let bobby: Person = table.get(2).data(&connection)?.unwrap();
assert_eq!(bobby.name, "Bobby");

pub fn delete<C: AsRef<SqliteConnection>>(
&self,
id: I,
connection: C
) -> SqliteResult<()>

Deletes an entry with the given primary key.

Example

// We insert 2 entries
table.insert(json!(1), &connection)?;
table.insert(json!(2), &connection)?;
let length = table.iter().id(&connection)?.len();
assert_eq!(length, 2);
// Remove the entry with id 1
table.delete(1, &connection)?;
// Table should only have entry now
let length = table.iter().id(&connection)?.len();
assert_eq!(length, 1);
// And we shouldn't be able to access the entry with primary key 1 now
assert!(table.get(1).id(&connection)?.is_none());