Trait nosqlite::Filter

pub trait Filter {
    fn where_<'a, A, B, C>(&self, _: &Iterator<'a, A, B, C>) -> Option<String>;

    fn and<B: Filter>(self, second: B) -> And<Self, B>
    where
        Self: Sized,
    { ... }

    fn or<B: Filter>(self, second: B) -> Or<Self, B>
    where
        Self: Sized,
    { ... }

    fn not(self) -> Not<Self>
    where
        Self: Sized,
    { ... }
}

Represents a condition which will determine what entries the operation can work on.

Required methods

fn where_<'a, A, B, C>(&self, _: &Iterator<'a, A, B, C>) -> Option<String>

Returns a string formatted for use in an SQL statement.

Loading content...
Provided methods

fn and<B: Filter>(self, second: B) -> And<Self, B> where
Self: Sized,

Allows chaining of multiple conditions.

fn or<B: Filter>(self, second: B) -> Or<Self, B> where
Self: Sized,

Allows another possible condition.

fn not(self) -> Not<Self> where
Self: Sized,

Negates the condition.

Example

table.insert(1, &connection)?;
table.insert(2, &connection)?;
table.insert(3, &connection)?;
table.insert(1, &connection)?;

// Use a filter that effectively is a not equal comparison
let numbers: Vec<u8> = table.iter().filter(field("").eq(1).not()).data(&connection)?;
assert_eq!(numbers.len(), 2);
assert!(!numbers.into_iter().any(|number| number == 1));

Loading content...