Trait nosqlite::FromSql

pub trait FromSql {
    fn column_result(value: ValueRef) -> Result<Self, FromSqlError>;
}

A trait for types that can be created from a SQLite value.

Note that FromSql and ToSql are defined for most integral types, but not u64 or usize. This is intentional; SQLite returns integers as signed 64-bit values, which cannot fully represent the range of these types. Rusqlite would have to decide how to handle negative values: return an error or reinterpret as a very large postive numbers, neither of which is guaranteed to be correct for everyone. Callers can work around this by fetching values as i64 and then doing the interpretation themselves or by defining a newtype and implementing FromSql/ToSql for it.

Required methods

fn column_result(value: ValueRef) -> Result<Self, FromSqlError>

Loading content...