Trait nosqlite::Key

pub trait Key {
    fn key(&self, data_key: &str) -> String;

    fn eq<T: Serialize>(self, value: T) -> Eq<Self, String>
    where
        Self: Sized,
    { ... }

    fn neq<T: Serialize>(self, value: T) -> Neq<Self, String>
    where
        Self: Sized,
    { ... }

    fn gt<T: Serialize>(self, value: T) -> Gt<Self, String>
    where
        Self: Sized,
    { ... }

    fn gte<T: Serialize>(self, value: T) -> Gte<Self, String>
    where
        Self: Sized,
    { ... }

    fn lt<T: Serialize>(self, value: T) -> Gt<String, Self>
    where
        Self: Sized,
    { ... }

    fn lte<T: Serialize>(self, value: T) -> Gte<String, Self>
    where
        Self: Sized,
    { ... }

    fn like<S: Display>(
        self, 
        matches_start: bool, 
        value: S, 
        matches_end: bool
    ) -> Like<Self, S>
    where
        Self: Sized,
    { ... }

    fn exists(self) -> Exists<Self>
    where
        Self: Sized,
    { ... }

    fn ascending(self) -> SortOrder<Self>
    where
        Self: Sized,
    { ... }

    fn descending(self) -> SortOrder<Self>
    where
        Self: Sized,
    { ... }
}

This can be used for filters or getting fields

Required methods

fn key(&self, data_key: &str) -> String

Produces the string that will be used by SQL.

Loading content...
Provided methods

fn eq<T: Serialize>(self, value: T) -> Eq<Self, String> where
Self: Sized,

Compares for equality.

Example

table.insert(json!({"number": 3}), &connection)?;
table.insert(json!({"number": 9}), &connection)?;
table.insert(json!({"number": 2}), &connection)?;
table.insert(json!({"number": 4}), &connection)?;
table.insert(json!({"number": 3}), &connection)?;
// Get only the entries where the number is equal to 3
let numbers: Vec<u8> = table.iter()
	.filter(field("number").eq(3)).field("number", &connection)?;
// Only 2 entries should've matched the filter
assert_eq!(numbers.len(), 2);
// They both should be equal to 3
assert_eq!(numbers.into_iter().any(|number| number != 3), false);

fn neq<T: Serialize>(self, value: T) -> Neq<Self, String> where
Self: Sized,

Compares for inequality.

Example

table.insert(json!({"number": 3}), &connection)?;
table.insert(json!({"number": 9}), &connection)?;
table.insert(json!({"number": 2}), &connection)?;
table.insert(json!({"number": 4}), &connection)?;
table.insert(json!({"number": 3}), &connection)?;
// Get only the entries where the number is not equal to 3
let numbers: Vec<u8> = table.iter()
	.filter(field("number").neq(3)).field("number", &connection)?;
// 3 entries should've matched the filter
assert_eq!(numbers.len(), 3);
// They both should be equal to 3
assert!(!numbers.into_iter().any(|number| number == 3));

fn gt<T: Serialize>(self, value: T) -> Gt<Self, String> where
Self: Sized,

Compares if it is greater than the value.

Example

table.insert(json!({"number": 3}), &connection)?;
table.insert(json!({"number": 9}), &connection)?;
table.insert(json!({"number": 2}), &connection)?;
table.insert(json!({"number": 4}), &connection)?;
table.insert(json!({"number": 3}), &connection)?;
// Get only the entries where the number is greater than 4
let numbers: Vec<u8> = table.iter()
	.filter(field("number").gt(4)).field("number", &connection)?;
// Only 9 is bigger than 4
assert_eq!(numbers.len(), 1);
assert!(numbers[0] > 4);

fn gte<T: Serialize>(self, value: T) -> Gte<Self, String> where
Self: Sized,

Compares if it is greater than or equal to the value.

Example

table.insert(json!({"number": 3}), &connection)?;
table.insert(json!({"number": 9}), &connection)?;
table.insert(json!({"number": 2}), &connection)?;
table.insert(json!({"number": 4}), &connection)?;
table.insert(json!({"number": 3}), &connection)?;
// Get only the entries where the number is greater than or equal to 4
let numbers: Vec<u8> = table.iter()
	.filter(field("number").gte(4)).field("number", &connection)?;
assert_eq!(numbers.len(), 2);
assert!(!numbers.into_iter().any(|number| number < 4));

fn lt<T: Serialize>(self, value: T) -> Gt<String, Self> where
Self: Sized,

Compares if it is less than the value.

Example

table.insert(json!({"number": 3}), &connection)?;
table.insert(json!({"number": 9}), &connection)?;
table.insert(json!({"number": 2}), &connection)?;
table.insert(json!({"number": 4}), &connection)?;
table.insert(json!({"number": 3}), &connection)?;
// Get only the entries where the number is less than 4
let numbers: Vec<u8> = table.iter()
	.filter(field("number").lt(4)).field("number", &connection)?;
assert_eq!(numbers.len(), 3);
assert!(!numbers.into_iter().any(|number| number >= 4));

fn lte<T: Serialize>(self, value: T) -> Gte<String, Self> where
Self: Sized,

Compares if it is greater than or equal to the variable.

Example

table.insert(json!({"number": 3}), &connection)?;
table.insert(json!({"number": 9}), &connection)?;
table.insert(json!({"number": 2}), &connection)?;
table.insert(json!({"number": 4}), &connection)?;
table.insert(json!({"number": 3}), &connection)?;
// Get only the entries where the number is less than or equal to 4
let numbers: Vec<u8> = table.iter()
	.filter(field("number").lte(4)).field("number", &connection)?;
assert_eq!(numbers.len(), 4);
assert!(!numbers.into_iter().any(|number| number > 4));

fn like<S: Display>(
self,
matches_start: bool,
value: S,
matches_end: bool
) -> Like<Self, S> where
Self: Sized,

Uses the SQL like comparison operator.

Example

table.insert(json!({"name": "Hiruna"}), &connection)?;
table.insert(json!({"name": "Alex"}), &connection)?;
table.insert(json!({"name": "Bob"}), &connection)?;
table.insert(json!({"name": "Haruna"}), &connection)?;
table.insert(json!({"name": "Felix"}), &connection)?;

// Get all names that end with 'x'
let names: Vec<String> = table.iter()
	.filter(field("name").like(true, 'x', false)).field("name", &connection)?;
// We should only match two names
assert_eq!(names.len(), 2);

// Get all names that start with `H` and end with 'runa'
let names: Vec<String> = table.iter()
	.filter(field("name").like(false, "H%runa", false)).field("name", &connection)?;
// We should only match two names
assert_eq!(names.len(), 2);

fn exists(self) -> Exists<Self> where
Self: Sized,

Whether the value exists in the JSON object and if it does exist, whether it is not null.

Example

table.insert(json!({"a": 3, "b": 20}), &connection)?;
table.insert(json!({"b": 3, "c": 20}), &connection)?;
table.insert(json!({"a": 89, "b": 3, "c": 20}), &connection)?;
table.insert(json!({"a": null, "b": 3, "c": 20}), &connection)?;

let ids: Vec<i64> = table.iter().filter(field("a").exists()).id(&connection)?;
// Only Entry 1 and 3 should have passed the filter
// Entry 2 doesn't have the `a` field and Entry 4 has an `a` field but it's `null`
// Therefore only two entries should have passed
assert_eq!(ids.len(), 2);
assert_eq!(ids[0], 1);
assert_eq!(ids[1], 3);

fn ascending(self) -> SortOrder<Self> where
Self: Sized,

This field is to be sorted in ascending order.

Example

table.insert(json!({"a": 4}), &connection)?;
table.insert(json!({"a": 8}), &connection)?;
table.insert(json!({"a": 6}), &connection)?;

let a: Vec<u8> = table.iter().sort(field("a").ascending()).field("a", &connection)?;

assert_eq!(a[0], 4);
assert_eq!(a[1], 6);
assert_eq!(a[2], 8);

fn descending(self) -> SortOrder<Self> where
Self: Sized,

This field is to be sorted in descending order.

Example

table.insert(json!({"a": 4}), &connection)?;
table.insert(json!({"a": 8}), &connection)?;
table.insert(json!({"a": 6}), &connection)?;

let a: Vec<u8> = table.iter().sort(field("a").descending()).field("a", &connection)?;

assert_eq!(a[0], 8);
assert_eq!(a[1], 6);
assert_eq!(a[2], 4);

Loading content...