Trait nosqlite::Sort

pub trait Sort {
    fn order_by(&self, data_key: &str) -> Vec<String>;

    fn and<B>(self, second: B) -> And<Self, B>
    where
        Self: Sized,
    { ... }
}

An expression that is used to sort the entries in the database before it is returned to the program.

Required methods

fn order_by(&self, data_key: &str) -> Vec<String>

Crates a list of SQL expressions to use for sorting

Normal users of this crate should not need to use this at all.

Loading content...
Provided methods

fn and<B>(self, second: B) -> And<Self, B> where
Self: Sized,

Add to the list of SQL expressions being used for sorting.

The second one will be taken into account after the first one.

Loading content...