Module nosqlite::util

A module for utility structs that don't do much on their own

Structs

And
A struct that represents AND.

Eq
A struct that represents equality.

Exists
A struct that checks whether a field exists and if that field is not null.

Gt
A struct that compares whether G > L.

Gte
A struct that compares whether G >= L.

Like
A struct that compares using the SQL LIKE comparison.

Neq
A struct that represents inequality.

Not
A struct that represents NOT.

Or
A struct that represents OR.

Enums

SortOrder
The order which the key will be sorted by